moved last routing functions
