+++ lib/route.js
@@ -3,7 +3,10 @@
 'use strict'
 
 const FindMyWay = require('find-my-way')
+const proxyAddr = require('proxy-addr')
 const Context = require('./context')
+const handleRequest = require('./handleRequest')
+const { hookRunner, hookIterator } = require('./hooks')
 const supportedMethods = ['DELETE', 'GET', 'HEAD', 'PATCH', 'POST', 'PUT', 'OPTIONS']
 const validation = require('./validation')
 const buildSchema = validation.build
@@ -30,21 +33,43 @@ module.exports = function buildRouting (options) {
   const schemaCache = new Map()
   schemaCache.put = schemaCache.set
 
-  let routeHandler
   let avvio
   let buildMiddie
   let fourOhFour
+  let trustProxy
+  let requestIdHeader
+  let querystringParser
+  let requestIdLogLabel
+  let logger
+  let hasLogger
+  let setupResponseListeners
+  let proxyFn
+  let modifyCoreObjects
+  let genReqId
+
+  let closing = false
 
   return {
     fill (args) {
-      routeHandler = args.routeHandler
       avvio = args.avvio
       buildMiddie = args.buildMiddie
       fourOhFour = args.fourOhFour
+      trustProxy = args.trustProxy
+      requestIdHeader = args.requestIdHeader
+      querystringParser = args.querystringParser
+      requestIdLogLabel = args.requestIdLogLabel
+      logger = args.logger
+      hasLogger = args.hasLogger
+      setupResponseListeners = args.setupResponseListeners
+      proxyFn = args.proxyFn
+      modifyCoreObjects = args.modifyCoreObjects
+      genReqId = args.genReqId
     },
     routing: router.lookup.bind(router), // router func to find the right handler to call
     route, // configure a route in the fastify instance
     prepareRoute,
+    routeHandler,
+    closeRoutes: () => { closing = true },
     printRoutes: router.prettyPrint.bind(router)
   }
 
@@ -227,6 +252,74 @@ module.exports = function buildRouting (options) {
   function throwIfAlreadyStarted (msg) {
     if (this[kState].started) throw new Error(msg)
   }
+
+  // HTTP request entry point, the routing has already been executed
+  function routeHandler (req, res, params, context) {
+    if (closing === true) {
+      const headers = {
+        'Content-Type': 'application/json',
+        'Content-Length': '80'
+      }
+      if (req.httpVersionMajor !== 2) {
+        headers.Connection = 'close'
+      }
+      res.writeHead(503, headers)
+      res.end('{"error":"Service Unavailable","message":"Service Unavailable","statusCode":503}')
+      if (req.httpVersionMajor !== 2) {
+        // This is not needed in HTTP/2
+        setImmediate(() => req.destroy())
+      }
+      return
+    }
+
+    req.id = req.headers[requestIdHeader] || genReqId(req)
+    req.originalUrl = req.url
+    var hostname = req.headers['host']
+    var ip = req.connection.remoteAddress
+    var ips
+
+    if (trustProxy) {
+      ip = proxyAddr(req, proxyFn)
+      ips = proxyAddr.all(req, proxyFn)
+      if (ip !== undefined && req.headers['x-forwarded-host']) {
+        hostname = req.headers['x-forwarded-host']
+      }
+    }
+
+    var childLogger = logger.child({ [requestIdLogLabel]: req.id, level: context.logLevel })
+
+    // added hostname, ip, and ips back to the Node req object to maintain backward compatibility
+    if (modifyCoreObjects) {
+      req.hostname = hostname
+      req.ip = ip
+      req.ips = ips
+
+      req.log = res.log = childLogger
+    }
+
+    childLogger.info({ req }, 'incoming request')
+
+    var queryPrefix = req.url.indexOf('?')
+    var query = querystringParser(queryPrefix > -1 ? req.url.slice(queryPrefix + 1) : '')
+    var request = new context.Request(params, req, query, req.headers, childLogger, ip, ips, hostname)
+    var reply = new context.Reply(res, context, request, childLogger)
+
+    if (hasLogger === true || context.onResponse !== null) {
+      setupResponseListeners(reply)
+    }
+
+    if (context.onRequest !== null) {
+      hookRunner(
+        context.onRequest,
+        hookIterator,
+        request,
+        reply,
+        middlewareCallback
+      )
+    } else {
+      middlewareCallback(null, request, reply)
+    }
+  }
 }
 
 // TODO replicated
@@ -236,3 +329,38 @@ function validateBodyLimitOption (bodyLimit) {
     throw new TypeError(`'bodyLimit' option must be an integer > 0. Got '${bodyLimit}'`)
   }
 }
+
+function middlewareCallback (err, request, reply) {
+  if (reply.sent === true) return
+  if (err != null) {
+    reply.send(err)
+    return
+  }
+
+  if (reply.context._middie !== null) {
+    reply.context._middie.run(request.raw, reply.res, reply)
+  } else {
+    onRunMiddlewares(null, null, null, reply)
+  }
+}
+
+function onRunMiddlewares (err, req, res, reply) {
+  if (err != null) {
+    reply.send(err)
+    return
+  }
+
+  if (reply.context.preParsing !== null) {
+    hookRunner(
+      reply.context.preParsing,
+      hookIterator,
+      reply.request,
+      reply,
+      handleRequest
+    )
+  } else {
+    handleRequest(null, reply.request, reply)
+  }
+}
+
+module.exports.onRunMiddlewares = onRunMiddlewares
